Update display inputs
