+++ assets/src/components/Configuration.js
@@ -42,7 +42,7 @@ import { LoopingLogo } from './utils/AnimatedLogo'
 
 import { BuildFragment } from './graphql/builds'
 import { LabelledInput } from './utils/LabelledInput'
-import { convertType } from './runbooks/Display'
+import { convertType } from './apps/app/runbooks/runbook/RunbookDisplay'
 import { DarkSelect } from './utils/Select'
 
 import { COMPONENT_LABEL } from './kubernetes/constants'
